See description below =D
Reverted to older splitversion, made splitversion config useful again, and added the splitMult config in place of MULT for more customizable physics

The new splitting is basically the same thing and has the same physics as the old one, the new one's code is just more complicated, so i reverted it. And the only thing that splitversion = 0 has use for is making the splitMult config useful (splitversion = 1 disables it because mult is already in use), only use splitversion = 0 if you want to customize your splitting more.

Pushback is still an issue, I haven't been able to fix that yet I need some help with it. So don't worry I haven't forgot about it.